Create WebglDrone.js
Javascript file that constructs the object, and adds functionality to the buttons displayed by the HTML file. This program is meant to simulate the flying of a drone around an object. the buttons allow one to control the camera on the Webgl element and "fly" the drone around the object